+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### ✨ Enhancements
 
+-   Changed the `filePath` argument in the MemoryViz CLI to be optional and added `--output` option.
+
 ### 🐛 Bug fixes
 
 -   Fixed issue where object boxes would be drawn on top of stack frames in diagrams with large left margins.
@@ -28,14 +30,12 @@ and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 -   Renamed the input for blank objects from `BLANK` to `.blank`.
 -   Created new type `.blank-frame` to denote blank stack frames.
 -   Replaced `seed` configuration option with general configuration option `roughjs_config`.
--   Changed the `filePath` argument in the MemoryViz CLI to be optional.
 
 ### ✨ Enhancements
 
 -   Created a CLI for MemoryViz.
 -   Added `--height` and `--width` options to MemoryViz CLI.
 -   Added `--roughjs_config` option to MemoryViz CLI.
--   Added options `--stdout`, and `--output` to MemoryViz CLI.
 
 ### 🐛 Bug fixes